Fix #397
New UI component Markdown editor, with onclick editing functionality and in sync UI with preview.

Closes #397
