+++ common/changes/@anticrm/presentation/fix-397_2021-11-12-09-34.json
@@ -1,0 +1,11 @@
+{
+  "changes": [
+    {
+      "comment": "MarkdownEditor with preview",
+      "type": "patch",
+      "packageName": "@anticrm/presentation"
+    }
+  ],
+  "packageName": "@anticrm/presentation",
+  "email": "haiodo@gmail.com"
+}
